crud category and sub category

+++ app/Http/Controllers/BrandController.php
@@ -40,14 +40,6 @@ class BrandController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     */
-    public function show(Item $item)
-    {
-        //
-    }
-
-    /**
      * Show the form for editing the specified resource.
      */
     public function edit(Brand $brand)
@@ -72,7 +64,10 @@ class BrandController extends Controller
      */
     public function destroy(Brand $brand)
     {
+        if (sizeOf($brand->items) > 0)
+            return redirect()->back()->with('message', 'Brand ' . $brand->nama . ' tidak Bisa dihapus masih ada item yang menggunakan brand ini.');
+
         $brand->delete();
-        return redirect('/admin/brand');
+        return redirect()->back()->with('message', 'Brand ' . $brand->nama . ' barhasil dihapus');
     }
 }
